DTO'er til alle model klasserne

+++ src/main/java/com/kinoxp/model/seat/Seat.java
@@ -7,7 +7,7 @@ import jakarta.persistence.*;
 public class Seat {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int seatId;
+    private Long seatId;
 
     @ManyToOne
     @JoinColumn(name = "theater_id")
@@ -18,18 +18,18 @@ public class Seat {
 
     public Seat() {}
 
-    public Seat(int seatId, Theater theater, int seatNumber, int rowNumber) {
+    public Seat(Long seatId, Theater theater, int seatNumber, int rowNumber) {
         this.seatId = seatId;
         this.theater = theater;
         this.seatNumber = seatNumber;
         this.rowNumber = rowNumber;
     }
 
-    public int getSeatId() {
+    public Long getSeatId() {
         return seatId;
     }
 
-    public void setSeatId(int seatId) {
+    public void setSeatId(Long seatId) {
         this.seatId = seatId;
     }
